doctest

+++ narwhals/series.py
@@ -682,7 +682,7 @@ class Series:
             We can pass any supported library such as Pandas, Polars, or PyArrow to `func`:
 
             >>> func(s_pd)
-            1.4724267269058975
+            np.float64(1.4724267269058975)
             >>> func(s_pl)
             1.4724267269058975
